Fixes #126 Refactor battle tick logic from processing.rs into battle/tick.rs
